修复部分BUG 20230517v2

+++ utills/index.js
@@ -1,8 +1,4 @@
 const fs = require("fs");
-const path = require("path");
-const { dirname } = require("path");
-
-const fsExtra = require("fs-extra");
 
 exports.errSend = (res, err, result, tip) => {
   if (err) {
@@ -17,25 +13,28 @@ exports.errSend = (res, err, result, tip) => {
 };
 
 // 将图片上传服务端，转化为图片路径，将图片路径上传到数据库
-exports.buffer = (req, res, avatar, pathMy) => {
-  const Key = "+爱的魔力转圈圈";
-  const path = "./public/user_pic/";
+exports.buffer = (req, res, avatar, pathMy, way = "no") => {
+  const Key = pathMy == "article_pic" ? "+1999100599999" : "+1999111299999";
+
+  const path = "./public/" + pathMy + "/";
 
   // 获取目标文件夹中的所有图片
   let files = fs.readdirSync(path);
 
-  // 删除替换前的图片
-  if (files.length > 0) {
-    files.map((v) => {
-      if (v.includes(req.user.id + Key)) {
-        fs.unlinkSync(path + v);
-      }
-    });
+  // 当不是新增文章时
+  if (way !== "add") {
+    // 删除替换前的图片
+    if (files.length > 0) {
+      files.map((v) => {
+        if (v.includes(req.user.id + Key)) {
+          fs.unlinkSync(path + v);
+        }
+      });
+    }
   }
 
-  //过滤data:URL
-
-  let base64url = avatar.url.replace(/^data:image\/\w+;base64,/, "");
+  //过滤data:URL、
+  let base64url = avatar["url"].replace(/^data:image\/\w+;base64,/, "");
   let dataBuffer = new Buffer.from(base64url, "base64");
 
   // 存储文件命名是使用当前时间，防止文件重名
